# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Complete eSIM platform with QPay payment integration and RoamWiFi API",
        "title": "eSIM Platform API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/analytics/products": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve product analytics data (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Analytics"
                ],
                "summary": "Get product analytics (Admin)",
                "responses": {
                    "200": {
                        "description": "Product analytics data",
                        "schema": {
                            "$ref": "#/definitions/handlers.ProductAnalyticsResponse"
                        }
                    }
                }
            }
        },
        "/admin/analytics/sales": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve sales analytics data (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Analytics"
                ],
                "summary": "Get sales analytics (Admin)",
                "responses": {
                    "200": {
                        "description": "Sales analytics data",
                        "schema": {
                            "$ref": "#/definitions/handlers.SalesAnalyticsResponse"
                        }
                    }
                }
            }
        },
...
        "/admin/packages/{priceId}/markup": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set or clear markup percent (clears override)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Packages"
                ],
                "summary": "Update package markup (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Provider Price ID",
                        "name": "priceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Markup payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdatePackageMarkupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/packages/{priceId}/override": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Set or clear override price (clears markup usage)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Packages"
                ],
                "summary": "Update package override price (Admin)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Provider Price ID",
                        "name": "priceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Override payload (null to clear)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdatePackageOverrideRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/pricing/exchange-rate": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Manually set the USD to MNT exchange rate (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Pricing"
                ],
                "summary": "Update exchange rate (Admin)",
                "parameters": [
                    {
                        "description": "Exchange rate",
                        "name": "rate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateExchangeRateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange rate updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/pricing/info": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get current pricing information including exchange rates and profit margins (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Pricing"
                ],
                "summary": "Get pricing information (Admin)",
                "responses": {
                    "200": {
                        "description": "Current pricing information",
                        "schema": {
                            "$ref": "#/definitions/handlers.PricingInfo"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/pricing/update-all": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Recalculate pricing for all products and packages (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Pricing"
                ],
                "summary": "Update all product pricing (Admin)",
                "responses": {
                    "200": {
                        "description": "All pricing updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/admin/products/sync": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Synchronize products from RoamWiFi API (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin",
                    "Products"
                ],
                "summary": "Sync products from RoamWiFi API (Admin)",
                "responses": {
                    "200": {
                        "description": "Products synced successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ProductAnalyticsResponse": {
            "type": "object",
            "properties": {
                "active_products": {
                    "type": "integer"
                },
                "inactive_products": {
                    "type": "integer"
                },
                "top_selling_products": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "total_products": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.SalesAnalyticsResponse": {
            "type": "object",
            "properties": {
                "average_order_value": {
                    "type": "number"
                },
                "completed_orders": {
                    "type": "integer"
                },
                "failed_orders": {
                    "type": "integer"
                },
                "pending_orders": {
                    "type": "integer"
                },
                "total_orders": {
                    "type": "integer"
                },
                "total_sales": {
                    "type": "number"
                }
            }
        },
...
        "handlers.UpdatePackageMarkupRequest": {
            "type": "object",
            "properties": {
                "markup_percent": {
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}